Finalize version of project 4
This is the final revision of Project 4 code. The code is fully functional however is not perfectly commented. This will form the basis for Project 5.

All code is organized into a hopefully readable form where each file has it's own purpose. Note: Interrupts MUST be in the main.c file in order to function. Otherwise, main.c is more of a header file that calls the appropriate functions.